requirements.txt
discord.py
peewee
requests
